+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - use image helper in uploader to resize image
 - fix for saving default value from admin and for backend is required validation
+### Fixed
+- fix radio buttons on Manage Options on boolean attribute
 
 ## [101.1.4] - 2022-11-19
 ### Fixed